Add files via upload
This is Saad in your account, Bojan. The change I did is that I made the navigation in the center. Also, I think it look better than better. What I am trying to say is that instead of having one big rectangle, we can have rectangle for each of them.